Move llm-related components in core into llm package
