+++ src/main/kotlin/org/jetbrains/research/testspark/tools/llm/generation/PromptManager.kt
@@ -15,12 +15,12 @@ import com.intellij.psi.util.PsiTypesUtil
 import org.jetbrains.research.testspark.bundles.TestSparkBundle
 import org.jetbrains.research.testspark.core.generation.importPattern
 import org.jetbrains.research.testspark.core.generation.packagePattern
-import org.jetbrains.research.testspark.core.generation.prompt.PromptGenerator
-import org.jetbrains.research.testspark.core.generation.prompt.configuration.ClassRepresentation
-import org.jetbrains.research.testspark.core.generation.prompt.configuration.MethodRepresentation
-import org.jetbrains.research.testspark.core.generation.prompt.configuration.PromptConfiguration
-import org.jetbrains.research.testspark.core.generation.prompt.configuration.PromptGenerationContext
-import org.jetbrains.research.testspark.core.generation.prompt.configuration.PromptTemplates
+import org.jetbrains.research.testspark.core.generation.llm.prompt.PromptGenerator
+import org.jetbrains.research.testspark.core.generation.llm.prompt.configuration.ClassRepresentation
+import org.jetbrains.research.testspark.core.generation.llm.prompt.configuration.MethodRepresentation
+import org.jetbrains.research.testspark.core.generation.llm.prompt.configuration.PromptConfiguration
+import org.jetbrains.research.testspark.core.generation.llm.prompt.configuration.PromptGenerationContext
+import org.jetbrains.research.testspark.core.generation.llm.prompt.configuration.PromptTemplates
 import org.jetbrains.research.testspark.data.CodeType
 import org.jetbrains.research.testspark.data.FragmentToTestData
 import org.jetbrains.research.testspark.helpers.generateMethodDescriptor
